+++ openidea/src/App.js
@@ -3,7 +3,7 @@ import logo from './logo.svg';
 import './App.css';
 import Navbar from './component/navbar'
 import ImageExampleFluid from './component/image'
-import IdeaCards from './component/ideaCards'
+import IdeaCards from './landing/ideaCards'
 class App extends Component {
   state = {
     ideas: [
